Move components into components' foldel
