+++ frontend/src/app/modules/landing/components/landing-content-item/landing-content-item.component.sass
@@ -1,4 +1,4 @@
-@import ../common-styles/variables
+@import ../../common-styles/variables
 .item-block
     display: grid
     grid-auto-flow: row
