+++ frontend/src/app/modules/landing/components/landing-content-item/landing-content-item.component.ts
