+++ frontend/src/app/modules/landing/components/landing-content/landing-content.component.sass
@@ -1,5 +1,5 @@
 @import src/styles/mixins
-@import ../common-styles/variables
+@import ../../common-styles/variables
     
 .content-container
     padding-top: $half-indent
