+++ frontend/src/app/modules/landing/components/landing-content/landing-content.component.ts
