+++ frontend/src/app/modules/landing/components/landing-demo/landing-demo.component.sass
@@ -1,5 +1,5 @@
 @import src/styles/mixins
-@import ../common-styles/variables
+@import ../../common-styles/variables
 
 .demo-continer
     background: url(https://kartinkin.com/uploads/posts/2021-04/1617607358_37-p-fon-tekhnicheskii-41.png) no-repeat
